+++ BHoM_Adapter/BHoM_Adapter.csproj
@@ -70,15 +70,19 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AdapterConfig.cs" />
+    <Compile Include="CRUD\Create.cs" />
     <Compile Include="CRUD\Delete.cs" />
     <Compile Include="CRUD\Read.cs" />
-    <Compile Include="CRUD\Replace.cs" />
-    <Compile Include="CRUD\UpdateOnly.cs" />
-    <Compile Include="HelperMethods\MapSpecialProperties.cs" />
+    <Compile Include="Extra\Extra.cs" />
+    <Compile Include="Push\Push.cs" />
+    <Compile Include="Push\Replace.cs" />
+    <Compile Include="CRUD\Update.cs" />
+    <Compile Include="Push\MapSpecialProperties.cs" />
     <Compile Include="BHoMAdapter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="CRUD\UpdateProperty.cs" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
